requirements.txt
pip
Flask
Cassandra-driver
request
jsonify
requests
requests_cache
json2html
